Reorganizes the Syntax classes into packages
