+++ src/main/scala/org/xalgorithms/rules/elements/TableSources.scala
@@ -1,0 +1,11 @@
+package org.xalgorithms.rules.elements
+
+class TableSource(val whens: Seq[When]) {
+}
+
+class ColumnTableSource(val name: String, val source: String, whens: Seq[When]) extends TableSource(whens) {
+}
+
+class ColumnsTableSource(val columns: Seq[String], whens: Seq[When]) extends TableSource(whens) {
+}
+
