+++ src/main/scala/org/xalgorithms/rules/elements/Values.scala
@@ -1,0 +1,14 @@
+package org.xalgorithms.rules.elements
+
+class Value {
+}
+
+class Reference(val section: String, val key: String) extends Value {
+}
+
+class Number(val value: Double) extends Value {
+}
+
+class StringValue(val value: String) extends Value {
+}
+
