+++ src/main/scala/org/xalgorithms/rules/steps/AssembleStep.scala
@@ -1,0 +1,7 @@
+package org.xalgorithms.rules.steps
+
+import org.xalgorithms.rules.elements.{ Column }
+
+class AssembleStep(val name: String, val columns: Seq[Column]) extends Step {
+}
+
