+++ src/main/scala/org/xalgorithms/rules/steps/AssignmentStep.scala
@@ -1,0 +1,7 @@
+package org.xalgorithms.rules.steps
+
+import org.xalgorithms.rules.elements.{ Assignment, Reference }
+
+class AssignmentStep(val table: Reference, val assignments: Seq[Assignment]) extends Step {
+}
+
