+++ src/main/scala/org/xalgorithms/rules/steps/FilterStep.scala
@@ -1,0 +1,7 @@
+package org.xalgorithms.rules.steps
+
+import org.xalgorithms.rules.elements.{ Reference, When }
+
+class FilterStep(val table: Reference, val filters: Seq[When]) extends Step {
+}
+
