+++ src/main/scala/org/xalgorithms/rules/steps/KeepStep.scala
@@ -1,0 +1,5 @@
+package org.xalgorithms.rules.steps
+
+class KeepStep(val name: String, val table: String) extends Step {
+}
+
