+++ src/main/scala/org/xalgorithms/rules/steps/MapStep.scala
@@ -1,0 +1,7 @@
+package org.xalgorithms.rules.steps
+
+import org.xalgorithms.rules.elements.{ Assignment, Reference }
+
+class MapStep(table: Reference, assignments: Seq[Assignment]) extends AssignmentStep(table, assignments) {
+}
+
