+++ src/main/scala/org/xalgorithms/rules/steps/ReduceStep.scala
@@ -1,0 +1,9 @@
+package org.xalgorithms.rules.steps
+
+import org.xalgorithms.rules.elements.{ Assignment, Reference, When }
+
+class ReduceStep(
+  val filters: Seq[When],
+  table: Reference, assignments: Seq[Assignment]) extends AssignmentStep(table, assignments) {
+}
+
