+++ src/main/scala/org/xalgorithms/rules/steps/Step.scala
@@ -1,0 +1,5 @@
+package org.xalgorithms.rules.steps
+
+class Step {
+}
+
